Plugins: port bootstrap, list and run plugins
This makes the changes necessary for Avocado-VT plugins to work
properly with the new style plugins code on Avocado.

About namespace and module installation, the Avocado-VT plugins now
live on their own namespace, avocado_vt, which will usually find it
itself installed on `/usr/lib/pythonX.Y/site-packages/avocado_vt`.

Avocado can find the plugins because Avocado-VT has plugins registered
in the correct entry points (avocado.plugins.cli and
avocado.plugins.cli.cmd).
