+++ avocado-plugins-vt.spec
@@ -43,7 +43,7 @@ Xunit output, among others.
 %dir /etc/avocado/conf.d
 %config(noreplace)/etc/avocado/conf.d/vt.conf
 %doc README.rst LICENSE
-%{python_sitelib}/avocado*
+%{python_sitelib}/avocado_vt*
 %{python_sitelib}/virttest*
 %{_datadir}/avocado-plugins-vt/backends/*
 %{_datadir}/avocado-plugins-vt/shared/*
